ambassador-jwt/simple-app/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy
Flask-Migrate
psycopg2
Flask-Script
pyjwt
cryptography
